last save in update penjualan

+++ resources/views/halamanutamatoko.blade.php
@@ -318,6 +318,7 @@
 					
 					
 						<li>{!! link_to(route('home.index'),"Tabel data")!!}</li>
+						<li>{!! link_to(route('all.index'),"Tabel data semua")!!}</li>
 				
 					</ul>
 				</div>
@@ -372,7 +373,7 @@
 						 </table>  
 						 <div class="pagination pagination-centered">
 						
-							{{$pelanggan->appends(['kategori' => $kategori->currentPage(),'produk' => $produk->currentPage(),'penjualan' => $penjualan->currentPage(),'detail' => $detail->currentPage()])->links()}}
+							{{$pelanggan->appends(['kategori' => $kategori->currentPage(),'produk' => $produk->currentPage(),'penjualan' => $penjualan->currentPage()])->links()}}
 							 </div>       
 							
                     </div>
@@ -430,7 +431,7 @@
 						 <div class="pagination pagination-centered">
 						
 		
-						 {{$kategori->appends(['pelanggan' => $pelanggan->currentPage(),'produk' => $produk->currentPage(),'penjualan' => $penjualan->currentPage(),'detail' => $detail->currentPage()])->links()}}
+						 {{$kategori->appends(['pelanggan' => $pelanggan->currentPage(),'produk' => $produk->currentPage(),'penjualan' => $penjualan->currentPage()])->links()}}
 
 
 							 </div>       
@@ -496,7 +497,7 @@
 							  </tbody>
 						 </table>  
 						 <div class="pagination pagination-centered">
-						 {{$produk->appends(['pelanggan' => $pelanggan->currentPage(),'kategori' => $kategori->currentPage(),'penjualan' => $penjualan->currentPage(),'detail' => $detail->currentPage()])->links()}}
+						 {{$produk->appends(['pelanggan' => $pelanggan->currentPage(),'kategori' => $kategori->currentPage(),'penjualan' => $penjualan->currentPage()])->links()}}
 
 
 							 </div>       
@@ -525,7 +526,7 @@
 <br>
 	</br>
 	<div class="row-fluid sortable">
-				<div class="box span6">
+				<div class="box span7">
 					<div class="box-header">
 						<h2><i class="halflings-icon white align-justify"></i><span class="break"></span>Tabel Penjualan</h2>
 			
@@ -537,6 +538,9 @@
 									  <th>Id Penjualan</th>
 									  <th>Nama Pelanggan</th>
 									  <th>Tanggal Pembelian</th>
+									  <th>Id Produk</th>
+									  <th>Jumlah</th>
+									  <th>Total</th>
 									  <th>Edit</th>                     
 									  <th>Delete</th>                     
 								  </tr>
@@ -544,29 +548,38 @@
 							  <tbody>
 							  
 							@foreach($penjualan as $tampil)
-							  
-								<tr>
-								<td>{!! $tampil->id!!}</td>
-							 <td>{!! $tampil->pelanggan->nama_pelanggan!!}</td>
-							 <td>{!! $tampil->tanggal_beli!!}</td>
-						    <td>{!! link_to(route('penjualan.edit',$tampil->id),'Edit') !!}</td>
+						       	@foreach($detail as $tampil2)
+								<tr> 	
+						@if($tampil2->id_penjualan==$tampil->id)
+						
+							    <td>{!! $tampil->id!!}</td>
+							    <td>{!! $tampil->pelanggan->nama_pelanggan!!}</td>
+							   <td>{!! $tampil->tanggal_beli!!}</td>
+							   <td>{!! $tampil2->produk->id!!}</td>
+							   <td>{!! $tampil2->jumlah!!}</td>
+							    <td>{!! $tampil2->total!!}</td>
+								<td>{!! link_to(route('penjualan.edit',$tampil->id),'Edit') !!}</td>
 				            <td>{!! Form::open(array('route'=>array('penjualan.destroy',$tampil->id),'method'=>'delete')) !!}
 					        {!! Form::submit('Delete',array("onclick"=>"return confirm('are you sure?')")) !!}
 					          {!! Form::close() !!}
 					          </td>
-		
+							 @else
+						
+							
 							 </tr>
-		                 	@endforeach
-							 <tr>
-							 <td>
-							 </tr>            
+							 @endif
+						  
+							 @endforeach
+							 @endforeach
+		                 	
+							
 							  </tbody>
 						 </table>  
 
 					
 						 <div class="pagination pagination-centered">
 						
-						 {{$penjualan->appends(['pelanggan' => $pelanggan->currentPage(),'kategori' => $kategori->currentPage(),'produk' => $produk->currentPage(),'detail' => $detail->currentPage()])->links()}}
+						 {{$penjualan->appends(['pelanggan' => $pelanggan->currentPage(),'kategori' => $kategori->currentPage(),'produk' => $produk->currentPage()])->links()}}
 							
 							 </div>      
                     </div>
@@ -581,72 +594,6 @@
 
 	</div>{{--/.fluid-container--}}
 	{!! link_to(route('penjualan.create'),"Tambah data penjualan")!!}
-
-
-
-<br>
-	</br>
-	<div class="row-fluid sortable">
-				<div class="box span6">
-					<div class="box-header">
-						<h2><i class="halflings-icon white align-justify"></i><span class="break"></span>Tabel Detail Penjualan</h2>
-			
-					</div>
-					<div class="box-content">
-						<table class="table">
-							  <thead>
-								  <tr>
-									  <th>Id Detail</th>
-									  <th>Id Penjualan</th>
-									  <th>Kode Produk</th>
-									  <th>Jumlah Pembelian</th>
-									  <th>Harga Total</th>
-									  <th>Edit</th>                     
-									  <th>Delete</th>                     
-								  </tr>
-							  </thead>   
-							  <tbody>
-							  
-							@foreach($detail as $tampil)
-							  
-								<tr>
-								<td>{!! $tampil->id!!}</td>
-							 <td>{!! $tampil->penjualan->id!!}</td>
-							 <td>{!! $tampil->produk->id!!}</td>
-							 <td>{!! $tampil->jumlah!!}</td>
-							 <td>{!! $tampil->total!!}</td>
-						    <td>{!! link_to(route('detail.edit',$tampil->id),'Edit') !!}</td>
-				            <td>{!! Form::open(array('route'=>array('detail.destroy',$tampil->id),'method'=>'delete')) !!}
-					        {!! Form::submit('Delete',array("onclick"=>"return confirm('are you sure?')")) !!}
-					          {!! Form::close() !!}
-					          </td>
-		
-							 </tr>
-		                 	@endforeach
-							 <tr>
-							 <td>
-							 </tr>            
-							  </tbody>
-						 </table>  
-
-					
-						 <div class="pagination pagination-centered">
-						
-						 {{$detail->appends(['pelanggan' => $pelanggan->currentPage(),'kategori' => $kategori->currentPage(),'produk' => $produk->currentPage(),'penjualan' => $penjualan->currentPage()])->links()}}
-							
-							 </div>      
-                    </div>
-				
-				</div>{{--/span--}}
-			
-		
-			
-
-		
-    
-
-	</div>{{--/.fluid-container--}}
-	{!! link_to(route('detail.create'),"Tambah data detail")!!}
 
 
 
@@ -744,3 +691,4 @@
 	
 </body>
 </html>
+
